Campos DECIMAL alterados para FLOAT

+++ MySQL/init-scripts/init.sql
@@ -3,8 +3,8 @@ USE Medidas;
 GRANT SELECT, INSERT ON Medidas.* TO 'app'@'%';
 CREATE TABLE Temperatura(
     Horario TIMESTAMP NOT NULL UNIQUE PRIMARY KEY COMMENT 'Horário de coleta das medidas.',
-    Temperatura DECIMAL(4, 2) COMMENT 'Temperatura medida em graus centígrados.',
-    Umidade DECIMAL(4, 2) UNSIGNED COMMENT 'Umidade relativa do ar em porcentagem.'
+    Temperatura FLOAT COMMENT 'Temperatura medida em graus centígrados.',
+    Umidade FLOAT UNSIGNED COMMENT 'Umidade relativa do ar em porcentagem.'
 );
